Things are starting to work again
ISSUE: ID was being lost when the theme was changed (solved) - Switched the forms for themes to POSTs
ISSUE: Variable admin_access was not initialized, causing problems when going to a page while not signed in (solved) - Made a check to see if the variable is set before it checks if it has a value.
Changed the navbar to have different links based on if the user is an admin. Navbar links to load_userpage.php if the user is not an admin. If they are, it links to load_page.php