Everything *SHOULD* work for the backend
All absolute paths were changed to local.

The session was being stored as a cookie and causing a lot of problems by not letting the variables be deleted. The session cookie now gets deleted to prevent that from happening.

When checking the login information against the database, the code was checking what the user put as the password against the database's password instead of a password stored within the database. It now queries for both the username and password at the same time and reports to the user if they were unable to login.

Themes were not being saved between pages. This was a result of my if statement checking if a button was pressed to change the theme. It now checks if a button was pressed to change the theme, then it checks what the session theme is, and if both those checks fail it sets the theme to the light theme.