Finished up the backend
Tested register to make sure it works properly, it does.
Added a register link to the navbar, as I found it annoying to go to login to get to register, so I figured users would find it annoying too.
Fixed some minor pathing errors.
Fixed an issue where when a non-admin user logged in they wouldn't be able to go to different pages because the drop downs weren't populated. Drop downs are now populated with the pages.